+++ src/main/java/lotto/LottoAplication.java
@@ -1,0 +1,7 @@
+package lotto;
+
+public class LottoAplication {
+    public static void main(String[] args) {
+
+    }
+}
