+++ src/main/java/lotto/model/NumbersGenerator.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class NumberGenerator {
+public class NumbersGenerator {
 
     public static List<Integer> autoGenerate() {
         List<Integer> numbers = IntStream.rangeClosed(1, 45)
