+++ src/test/java/lotto/LottoRetailerTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.model.LottoRetailer;
+import lotto.model.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
