+++ src/test/java/lotto/OrderTest.java
@@ -1,11 +1,11 @@
 package lotto;
 
+import lotto.model.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class OrderTest {
